Fix build: move CGR_MAX_ROUTES macros to header, fix shift overflow
- CGR_MAX_ROUTES_PI/ESP were #ifndef-guarded in cgr.c, invisible to
  forwarding.c. Moved to cgr.h so callers can use them.
- (uint64_t) cast on self_id << 32 to fix undefined left-shift on
  a 32-bit type.
- Silence read() warn_unused_result in main_pi.c.

+++ node/forwarding.c
@@ -43,7 +43,7 @@ uint64_t node_generate_bundle(node_state_t *ns, node_id_t dst,
     memset(&b, 0, sizeof(b));
     /* Bundle id: combine self_id + a counter. For MVP use a static counter. */
     static uint64_t counter = 1;
-    b.bundle_id    = (ns->self_id << 32) | (counter++);
+    b.bundle_id    = ((uint64_t)ns->self_id << 32) | (counter++);
     b.src          = ns->self_id;
     b.dst          = dst;
     b.curr         = ns->self_id;
